Create repo if it doesnt exist in the database
