+++ api/README.md
@@ -14,13 +14,13 @@ cp .env.dist .env.dev
 
 ```bash
 # development
-$ npm run start
+$ yarn start
 
 # watch mode
-$ npm run start:dev
+$ yarn start:dev
 
 # production mode
-$ npm run start:prod
+$ yarn start:prod
 ```
 
 ### With Docker
