+++ api/src/github/github.service.ts
@@ -1,9 +1,14 @@
 import { Injectable, HttpService } from '@nestjs/common';
 
-interface GithubRepoData {
+interface IDependenciesData {
   name: string;
   path: string;
   branch: string;
+  token: string;
+}
+
+interface IRepoData {
+  fullName: string;
   token: string;
 }
 
@@ -11,7 +16,7 @@ interface GithubRepoData {
 export class GithubService {
   constructor(private readonly httpService: HttpService) {}
 
-  async getPackageJson(data: GithubRepoData) {
+  async getPackageJson(data: IDependenciesData) {
     return this.httpService
       .get(
         `https://api.github.com/repos/${data.name}/contents/${data.path}package.json?ref=${data.branch}`,
@@ -25,7 +30,7 @@ export class GithubService {
       .toPromise();
   }
 
-  async getYarnLock(data: GithubRepoData) {
+  async getYarnLock(data: IDependenciesData) {
     return this.httpService
       .get(
         `https://api.github.com/repos/${data.name}/contents/${data.path}yarn.lock?ref=${data.branch}`,
@@ -36,6 +41,17 @@ export class GithubService {
           },
         },
       )
+      .toPromise();
+  }
+
+  async getRepository(data: IRepoData) {
+    return this.httpService
+      .get(`https://api.github.com/repos/${data.fullName}`, {
+        headers: {
+          Authorization: `token ${data.token}`,
+          Accept: 'application/vnd.github.machine-man-preview+json',
+        },
+      })
       .toPromise();
   }
 }
