+++ api/src/repository/repository.module.ts
@@ -6,9 +6,7 @@ import { Repository } from './repository.entity';
 import { DependenciesQueue } from '../queue/dependencies.queue';
 import { QueueModule } from '../queue/queue.module';
 import { GithubModule } from '../github/github.module';
-import { GithubService } from '../github/github.service';
 import { UsersModule } from '../users/users.module';
-import { UsersService } from '../users/users.service';
 
 @Module({
   imports: [
