--- front/src/containers/AddRepository/index.ts
@@ -1,1 +1,0 @@
-export { default } from './AddRepository'
