--- front/src/containers/RepoContent/index.ts
@@ -1,1 +1,0 @@
-export { default } from './RepoContent'
